# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WhatsApp Web Panel API",
    "version": "1.0.0",
    "description": "Cookie tabanli oturum (express-session) + CSRF korumali API. /api altindaki mutating (POST/PUT/DELETE) endpoint'ler icin CSRF header gonderilmelidir."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "whatsapp.sid"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "xsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-XSRF-TOKEN"
      }
    },
    "parameters": {
      "AccountIdHeader": {
        "name": "X-Account-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string" },
        "description": "Multi-account icin aktif hesabi secmekte kullanilir."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "requestId": { "type": "string" }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/webhooks": {
      "get": {
        "tags": ["Webhooks"],
        "summary": "Webhooks list (admin)",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "responses": {
          "200": { "description": "Webhooks", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["Webhooks"],
        "summary": "Webhook create (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "name": { "type": "string" },
                  "url": { "type": "string" },
                  "events": { "type": "string" },
                  "is_active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Created", "content": { "application/json": { "schema": { "type": "object" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/webhooks/{id}": {
      "put": {
        "tags": ["Webhooks"],
        "summary": "Webhook update (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      },
      "delete": {
        "tags": ["Webhooks"],
        "summary": "Webhook delete (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
...
    "/api/templates": {
      "get": {
        "tags": ["Templates"],
        "summary": "Templates list",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "responses": { "200": { "description": "Templates", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } } }
      },
      "post": {
        "tags": ["Templates"],
        "summary": "Template create (admin/manager)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "content"],
                "properties": {
                  "name": { "type": "string" },
                  "content": { "type": "string" },
                  "variables": { "type": ["string", "array", "null"], "items": { "type": "string" } },
                  "category": { "type": ["string", "null"] }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Created", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
    "/api/templates/{id}": {
      "get": {
        "tags": ["Templates"],
        "summary": "Template get",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Template", "content": { "application/json": { "schema": { "type": "object" } } } }, "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      },
      "put": {
        "tags": ["Templates"],
        "summary": "Template update (admin/manager)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name", "content"], "properties": { "name": { "type": "string" }, "content": { "type": "string" }, "variables": { "type": ["string", "array", "null"], "items": { "type": "string" } }, "category": { "type": ["string", "null"] } } } } } },
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      },
      "delete": {
        "tags": ["Templates"],
        "summary": "Template delete (admin/manager)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Users list (admin)",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "responses": { "200": { "description": "Users", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } } }
      },
      "post": {
        "tags": ["Users"],
        "summary": "User create (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password", "roleId"],
                "properties": {
                  "username": { "type": "string" },
                  "display_name": { "type": "string" },
                  "password": { "type": "string" },
                  "roleId": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Created", "content": { "application/json": { "schema": { "type": "object" } } } }, "409": { "description": "Username exists", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      }
    },
    "/api/users/{id}": {
      "delete": {
        "tags": ["Users"],
        "summary": "User delete (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/api/users/{id}/role": {
      "put": {
        "tags": ["Users"],
        "summary": "User role update (admin)",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/AccountIdHeader" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["roleId"], "properties": { "roleId": { "type": "integer" } } } } } },
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/api/users/me/preferences": {
      "put": {
        "tags": ["Users"],
        "summary": "My preferences update",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AccountIdHeader" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
...
  }
}